refactor and reorganise collection sort
